+++ core/src/main/java/org/apache/calcite/util/BuiltInMethod.java
@@ -342,6 +342,8 @@ public enum BuiltInMethod {
   STRING_TO_DATE(DateTimeUtils.class, "dateStringToUnixDate", String.class),
   STRING_TO_TIME(DateTimeUtils.class, "timeStringToUnixDate", String.class),
   STRING_TO_TIMESTAMP(DateTimeUtils.class, "timestampStringToUnixDate", String.class),
+  REGEX_STRING_TO_TIMESTAMP(SqlFunctions.class, "regexStringToTimestamp",
+      String.class, String.class),
   STRING_TO_TIME_WITH_LOCAL_TIME_ZONE(SqlFunctions.class, "toTimeWithLocalTimeZone",
       String.class),
   TIME_STRING_TO_TIME_WITH_LOCAL_TIME_ZONE(SqlFunctions.class, "toTimeWithLocalTimeZone",